# openapi.json
{
  "components": {
    "schemas": {
...
      "CountryCreate": {
        "additionalProperties": false,
        "description": "Request body for creating a trip country.",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CountryCreate",
        "type": "object"
      },
      "CountryEnvelope": {
        "description": "`{ \"country\": Country }`.",
        "properties": {
          "country": {
            "$ref": "#/components/schemas/CountryOut"
          }
        },
        "required": [
          "country"
        ],
        "title": "CountryEnvelope",
        "type": "object"
      },
      "CountryListEnvelope": {
        "description": "`{ \"countries\": [Country] }`.",
        "properties": {
          "countries": {
            "items": {
              "$ref": "#/components/schemas/CountryOut"
            },
            "title": "Countries",
            "type": "array"
          }
        },
        "required": [
          "countries"
        ],
        "title": "CountryListEnvelope",
        "type": "object"
      },
      "CountryOut": {
        "description": "Wire representation of a trip country.",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "flag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flag"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "item_count": {
            "title": "Item Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "code",
          "flag",
          "is_default",
          "sort_order",
          "item_count"
        ],
        "title": "CountryOut",
        "type": "object"
      },
      "CountryUpdate": {
        "additionalProperties": false,
        "description": "Request body for partially updating a trip country.",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          }
        },
        "title": "CountryUpdate",
        "type": "object"
      },
      "CurrencyCreate": {
        "additionalProperties": false,
        "description": "Request body for creating a trip currency.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "is_primary": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Primary"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "code"
        ],
        "title": "CurrencyCreate",
        "type": "object"
      },
      "CurrencyEnvelope": {
        "description": "`{ \"currency\": Currency }`.",
        "properties": {
          "currency": {
            "$ref": "#/components/schemas/CurrencyOut"
          }
        },
        "required": [
          "currency"
        ],
        "title": "CurrencyEnvelope",
        "type": "object"
      },
      "CurrencyListEnvelope": {
        "description": "`{ \"currencies\": [Currency] }`.",
        "properties": {
          "currencies": {
            "items": {
              "$ref": "#/components/schemas/CurrencyOut"
            },
            "title": "Currencies",
            "type": "array"
          }
        },
        "required": [
          "currencies"
        ],
        "title": "CurrencyListEnvelope",
        "type": "object"
      },
      "CurrencyOut": {
        "description": "Wire representation of a trip currency.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_primary": {
            "title": "Is Primary",
            "type": "boolean"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "id",
          "code",
          "symbol",
          "is_primary",
          "sort_order"
        ],
        "title": "CurrencyOut",
        "type": "object"
      },
      "CurrencyUpdate": {
        "additionalProperties": false,
        "description": "Request body for partially updating a trip currency.",
        "properties": {
          "is_primary": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Primary"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "title": "CurrencyUpdate",
        "type": "object"
      },
      "ErrorBodyOut": {
        "description": "The body of an error: a top-level code, its params, and optional per-field detail.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/FieldErrorOut"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorBodyOut",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "`{ \"error\": { \"code\": ..., \"params\": {...}, \"fields\": {...}? } }`.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyOut"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "FieldErrorOut": {
        "description": "One field's `{code, params}` failure.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "FieldErrorOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Expense tracking for a shared trip: people, currencies, countries and labels belong to a trip; every line item records who paid and how the cost splits across the roster.\n\nJSON in, JSON out. Every response is a single object with its payload under a named key, never a bare array. Amounts are plain numbers in the currency's own units - typed values to 2 fraction digits, computed shares to 6 - and are taken on input as canonical decimal strings (`\"18400.50\"`). Nothing is ever converted between currencies: balances and statistics are per currency.\n\nErrors are `{code, params}` with no message text, so a client renders them through its own catalog in its own language.",
    "title": "Trip Accounter API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/trips/{slug}/countries": {
      "get": {
        "description": "List a trip's countries with their item counts.",
        "operationId": "list_countries_api_v1_trips__slug__countries_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Countries",
        "tags": [
          "countries"
        ]
      },
      "post": {
        "description": "Add a new country to a trip.",
        "operationId": "create_country_api_v1_trips__slug__countries_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Country",
        "tags": [
          "countries"
        ]
      }
    },
    "/api/v1/trips/{slug}/countries/{country_id}": {
      "delete": {
        "description": "Delete a country from a trip.",
        "operationId": "delete_country_api_v1_trips__slug__countries__country_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "country_id",
            "required": true,
            "schema": {
              "title": "Country Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Delete Country",
        "tags": [
          "countries"
        ]
      },
      "patch": {
        "description": "Update a trip country's fields.",
        "operationId": "update_country_api_v1_trips__slug__countries__country_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "country_id",
            "required": true,
            "schema": {
              "title": "Country Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Update Country",
        "tags": [
          "countries"
        ]
      }
    },
    "/api/v1/trips/{slug}/currencies": {
      "get": {
        "description": "List a trip's currencies, in sort order.",
        "operationId": "list_currencies_api_v1_trips__slug__currencies_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Currencies",
        "tags": [
          "currencies"
        ]
      },
      "post": {
        "description": "Add a new currency to a trip.",
        "operationId": "create_currency_api_v1_trips__slug__currencies_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Currency",
        "tags": [
          "currencies"
        ]
      }
    },
    "/api/v1/trips/{slug}/currencies/{currency_id}": {
      "delete": {
        "description": "Delete a currency from a trip.",
        "operationId": "delete_currency_api_v1_trips__slug__currencies__currency_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "currency_id",
            "required": true,
            "schema": {
              "title": "Currency Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Delete Currency",
        "tags": [
          "currencies"
        ]
      },
      "patch": {
        "description": "Update a trip currency's fields.",
        "operationId": "update_currency_api_v1_trips__slug__currencies__currency_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "currency_id",
            "required": true,
            "schema": {
              "title": "Currency Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Update Currency",
        "tags": [
          "currencies"
        ]
      }
    },
...
  }
}